revised steering code

+++ Embedded/JetsonTX2/withoutROS/serial_arduino/msq_serial.cpp
@@ -3,7 +3,7 @@
 
 int fd, rdcnt;
 int msqid;
-int buffer;
+float buffer;
 
 key_t key = 54321;
 struct message recv_msg;
